+++ src/Nethermind/Nethermind.Db/IFlatDbConfig.cs
@@ -8,13 +8,13 @@ namespace Nethermind.Db;
 public interface IFlatDbConfig : IConfig
 {
     [ConfigItem(Description = "Block cache size budget", DefaultValue = "1073741824")]
-    long BlockCacheSizeBudget { get; set; }
+    ulong BlockCacheSizeBudget { get; set; }
 
     [ConfigItem(Description = "Fixed compaction schedule offset in blocks. When 0 or greater, overrides the per-instance offset in the metadata DB, which is neither read nor updated. Only the value modulo CompactSize matters. -1 to use the stored offset, generating a random one when absent.", DefaultValue = "-1")]
     long CompactionOffset { get; set; }
 
     [ConfigItem(Description = "Compact size", DefaultValue = "32")]
-    int CompactSize { get; set; }
+    ulong CompactSize { get; set; }
 
     [ConfigItem(Description = "Enabled", DefaultValue = "false")]
     bool Enabled { get; set; }
@@ -35,10 +35,10 @@ public interface IFlatDbConfig : IConfig
     int MaxInFlightCompactJob { get; set; }
 
     [ConfigItem(Description = "Max reorg depth — the force-persist backstop used when EnableLongFinality is off: once the in-memory depth exceeds it while finality is stalled, persistence is forced to bound memory.", DefaultValue = "256")]
-    int MaxReorgDepth { get; set; }
+    ulong MaxReorgDepth { get; set; }
 
     [ConfigItem(Description = "Minimum reorg depth", DefaultValue = "128")]
-    int MinReorgDepth { get; set; }
+    ulong MinReorgDepth { get; set; }
 
     [ConfigItem(Description = "Lower bound, in bytes, for the RocksDB write buffer (memtable) size of the flat-state columns. The per-batch adjuster never shrinks a column's memtable below this value. Raising it lets frequent small persistence batches (small CompactSize) coalesce and deduplicate in the memtable instead of churning L0, decoupling write amplification from CompactSize.", DefaultValue = "16777216")]
     long PersistenceWriteBufferFloor { get; set; }
@@ -47,7 +47,7 @@ public interface IFlatDbConfig : IConfig
     bool RegenerateCompactionOffset { get; set; }
 
     [ConfigItem(Description = "Trie cache memory target", DefaultValue = "536870912")]
-    long TrieCacheMemoryBudget { get; set; }
+    ulong TrieCacheMemoryBudget { get; set; }
 
     [ConfigItem(Description = "Trie warmer worker count (-1 for processor count - 1, 0 to disable)", DefaultValue = "-1")]
     int TrieWarmerWorkerCount { get; set; }
@@ -59,7 +59,7 @@ public interface IFlatDbConfig : IConfig
     bool EnableLongFinality { get; set; }
 
     [ConfigItem(Description = "Force-persist backstop used when EnableLongFinality is on, in place of MaxReorgDepth. The persisted-snapshot tier serves deep reorgs, so this is much larger than the non-long-finality backstop.", DefaultValue = "90000")]
-    int LongFinalityMaxReorgDepth { get; set; }
+    ulong LongFinalityMaxReorgDepth { get; set; }
 
     [ConfigItem(Description = "Maximum number of in-memory base snapshots before conversion to the persisted-snapshot tier kicks in. Counted as `SnapshotCount` of the in-memory repository, not a block-distance depth.", DefaultValue = "128")]
     int MaxInMemoryBaseSnapshotCount { get; set; }
@@ -77,7 +77,7 @@ public interface IFlatDbConfig : IConfig
     bool PersistedSnapshotPunchHoleOnReclaim { get; set; }
 
     [ConfigItem(Description = "Max persisted snapshot compaction size (hierarchical compaction ceiling for persisted layer), in blocks", DefaultValue = "1048576")]
-    int PersistedSnapshotMaxCompactSize { get; set; }
+    ulong PersistedSnapshotMaxCompactSize { get; set; }
 
     [ConfigItem(Description = "Validate persisted snapshots against in-memory snapshots after conversion (debug/diagnostic only)", DefaultValue = "false")]
     bool ValidatePersistedSnapshot { get; set; }
